service-deviceregistry/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking
#
asgiref==3.7.2
    # via django
django==4.2.1
    # via
    #   -r requirements.in
    #   djangorestframework
django-rest-framework==0.1.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via django-rest-framework
iniconfig==2.0.0
    # via pytest
packaging==23.1
    # via pytest
pillow==9.5.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.6
    # via -r requirements.in
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via
    #   -r requirements.in
    #   django
unittest-parametrize==1.0.0
    # via -r requirements.in
